Made listener priorities bytes to avoid confusion
See changelog for more details

+++ api/src/main/kotlin/org/kryptonmc/krypton/api/event/Listener.kt
@@ -8,4 +8,4 @@ package org.kryptonmc.krypton.api.event
  * @author Callum Seabrook
  */
 @Target(AnnotationTarget.FUNCTION)
-annotation class Listener(val priority: Int = ListenerPriority.MEDIUM)
+annotation class Listener(val priority: Byte = ListenerPriority.MEDIUM)
